fix ColorValueBounds

+++ Nez.Portable/Utils/Bounds/ColorValueBounds.cs
@@ -14,10 +14,10 @@ namespace Nez
 
         public Color nextValue()
         {
-            var nR = min.R + (max.R - min.R) * Random.nextFloat();
-            var nG = min.G + (max.G - min.G) * Random.nextFloat();
-            var nB = min.B + (max.B - min.B) * Random.nextFloat();
-            var nA = min.A + (max.A - min.A) * Random.nextFloat();
+            var nR = (int) (min.R + (max.R - min.R) * Random.nextFloat());
+            var nG = (int) (min.G + (max.G - min.G) * Random.nextFloat());
+            var nB = (int) (min.B + (max.B - min.B) * Random.nextFloat());
+            var nA = (int) (min.A + (max.A - min.A) * Random.nextFloat());
 
             return new Color(nR, nG, nB, nA);
         }
